requirements.txt
tensorflow>=2.13.0
keras>=2.13.0
streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=10.0.0
scikit-learn>=1.3.0
openvino>=2023.0.0
openvino-dev>=2023.0.0

# Optional dependencies (install separately if needed)
# scikit-image>=0.21.0  # Optional - may have issues with Python 3.13 on Windows
# Note: If you need scikit-image, try: pip install scikit-image --no-build-isolation
# Or use pre-built wheel: pip install scikit-image --only-binary :all:

# Note: Tkinter usually comes pre-installed with Python
# If not available, install: sudo apt-get install python3-tk (Linux)
# or it's included in Python installer (Windows/Mac)